phases_articles_retrieval/requirements.txt


python-dotenv
requests
biopython
PyQt5
googlesearch-python
scholarly
